refactor(web): extract community messaging presentation
