+++ src/web/src/components/community/channels/forum-view.tsx
@@ -17,7 +17,7 @@ import { ConfirmDialog } from "@/components/ui/confirm-dialog"
 import { tid } from "@/lib/community/testids"
 import type { ForumThread } from "@/lib/community/models/message"
 import type { Member } from "@/lib/community/models/people"
-import { VirtualRows } from "../messages/virtual-cursor-list"
+import { VirtualRows } from "@/components/ui/virtual-rows"
 import { useVirtualCursorSentinel } from "@/hooks/community/use-virtual-cursor-sentinel"
 import { tagColorClassName, tagColorStyle } from "@/lib/community/tag-color"
 import { cn } from "@/lib/utils"
