+++ src/web/src/modules/community/client/messaging/index.ts
@@ -1,0 +1,3 @@
+export { MessageList } from "./message-list"
+export { MessageRow } from "./message-row"
+export { TypingIndicator } from "./typing-indicator"
