+++ src/web/src/modules/community/client/messaging/internal/message-list-view.tsx
@@ -3,9 +3,9 @@ import { ArrowDown, ImageIcon, X } from "lucide-react"
 import { Button } from "@/components/ui/button"
 import { Skeleton } from "@/components/ui/skeleton"
 import { NumberTicker } from "@/components/ui/number-ticker"
-import { ChannelIcon } from "../channels/channel-icon"
-import { MessageShareDialog } from "./message-share-dialog"
-import { TypingIndicator } from "./typing-indicator"
+import { ChannelIcon } from "@/components/community/channels/channel-icon"
+import { MessageShareDialog } from "@/components/community/messages/message-share-dialog"
+import { TypingIndicator } from "../typing-indicator"
 import type { MessageListController } from "./message-list-controller"
 import type { ResolvedMessageListProps } from "./message-list-types"
